refactor: sharedPreference를 사용하는 곳에서 관리하도록 수정
